Allow multiple regularizers of identical operator
When adding a regularizer by name, only everything up to the first
whitespace in the string will be considered to match the name of an
operator function. This allows adding multiple instances of the same
operator, when passing a dict to Optimizer.set_regularizers.
This is useful for operators with more parameters that may be used in
multiple different ways, e.g. the axis parameter of operators.smooth.